pyproject.toml
[project]
name = "chcode"
version = "0.1.0"
description = "Terminal-based AI coding agent with typer+rich"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "questionary>=2.1.0",
    "prompt-toolkit>=3.0.0",
    "aiosqlite>=0.22.1",
    "bs4>=0.0.2",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "tavily>=1.1.0",
    "charset-normalizer>=3.4.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    "markdownify>=0.14.0",
    "aiofiles>=25.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
chcode = "chcode.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --timeout=30"

[tool.coverage.run]
source = ["chcode"]
omit = ["chcode/__main__.py"]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "pass",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
